elbowMethodDemonstration
Given a data set, it plots the un-clustered data set and runs the elbow method to determine the optimal amount of clusters for the data set. It does this by calculating the average distance from each point to the center of the assigned cluster. It then selects the point where the slope's change is below the total average (which is the point at which the change between k values is marginal). It then finds this optimal k value and plots the elbow plot with this optimal k value indicated with a red dashed line. It then displays a plot of the data set sorted by k means with the number of clusters determined  by the elbow method.